# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "PuntosFieles API",
    "version": "1.3.7",
    "description": "Route-derived OpenAPI document for versioned /api/v1 routes, with curated request/response metadata layered onto the discovered route surface."
  },
  "servers": [
    {
      "url": "https://api.puntos.local"
    }
  ],
  "paths": {
...
    "/api/v1/admin/staff": {
      "get": {
        "summary": "List staff users for the current business",
        "responses": {
          "200": {
            "description": "Staff list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StaffListResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Role not allowed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "staffAuth": []
          }
        ],
        "x-tenant-context": true,
        "x-required-roles": [
          "OWNER"
        ],
        "x-plan-features": [
          "staff_management"
        ]
      },
      "post": {
        "summary": "Create a staff user",
        "responses": {
          "200": {
            "description": "Staff user created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StaffDetailResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Role not allowed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Email already registered",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "staffAuth": []
          }
        ],
        "x-csrf-required": true,
        "x-tenant-context": true,
        "x-required-roles": [
          "OWNER"
        ],
        "x-plan-features": [
          "staff_management"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StaffCreateRequest"
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/staff/{id}": {
      "patch": {
        "summary": "Update a staff user",
        "responses": {
          "200": {
            "description": "Staff user updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StaffDetailResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Role not allowed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Staff user not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "staffAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "x-csrf-required": true,
        "x-tenant-context": true,
        "x-required-roles": [
          "OWNER"
        ],
        "x-plan-features": [
          "staff_management"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StaffUpdateRequest"
              }
            }
          }
        }
      }
    },
...
    "/api/v1/staff/award": {
      "post": {
        "summary": "Award points to a customer",
        "responses": {
          "200": {
            "description": "Points awarded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StaffAwardResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Permission denied",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "staffAuth": []
          }
        ],
        "x-csrf-required": true,
        "x-tenant-context": true,
        "x-required-permissions": [
          "staff.award"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AwardPointsRequest"
              }
            }
          }
        }
      }
    },
...
    "/api/v1/staff/login": {
      "post": {
        "summary": "Staff login",
        "responses": {
          "200": {
            "description": "Authenticated; cookie __Host-pf_staff is set"
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StaffLoginRequest"
              }
            }
          }
        }
      }
    },
    "/api/v1/staff/logout": {
      "post": {
        "summary": "Derived POST /api/v1/staff/logout",
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "security": [
          {
            "staffAuth": []
          }
        ],
        "x-csrf-required": true
      }
    },
    "/api/v1/staff/me": {
      "get": {
        "summary": "Current staff session",
        "responses": {
          "200": {
            "description": "Current staff session"
          },
          "401": {
            "description": "Not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "staffAuth": []
          }
        ],
        "x-tenant-context": true
      }
    },
    "/api/v1/staff/permissions": {
      "get": {
        "summary": "Derived GET /api/v1/staff/permissions",
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "security": [
          {
            "staffAuth": []
          }
        ],
        "x-tenant-context": true
      }
    },
    "/api/v1/staff/program": {
      "get": {
        "summary": "Derived GET /api/v1/staff/program",
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "security": [
          {
            "staffAuth": []
          }
        ],
        "x-tenant-context": true
      }
    },
    "/api/v1/staff/redeem": {
      "post": {
        "summary": "Derived POST /api/v1/staff/redeem",
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "security": [
          {
            "staffAuth": []
          }
        ],
        "x-csrf-required": true,
        "x-tenant-context": true,
        "x-required-permissions": [
          "staff.redeem"
        ],
        "x-plan-features": [
          "redemptions"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StaffRedeemRequest"
              }
            }
          }
        }
      }
    },
    "/api/v1/staff/refund": {
      "post": {
        "summary": "Derived POST /api/v1/staff/refund",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StaffRefundResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "staffAuth": []
          }
        ],
        "x-csrf-required": true,
        "x-tenant-context": true,
        "x-required-permissions": [
          "staff.refund"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StaffRefundRequest"
              }
            }
          }
        }
      }
    },
    "/api/v1/staff/rewards": {
      "get": {
        "summary": "Derived GET /api/v1/staff/rewards",
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "security": [
          {
            "staffAuth": []
          }
        ],
        "x-tenant-context": true,
        "x-plan-features": [
          "rewards"
        ]
      }
    },
    "/api/v1/staff/settle-pending": {
      "post": {
        "summary": "Derived POST /api/v1/staff/settle-pending",
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "security": [
          {
            "staffAuth": []
          }
        ],
        "x-csrf-required": true,
        "x-tenant-context": true,
        "x-required-permissions": [
          "staff.refund"
        ],
        "x-plan-features": [
          "lifecycle_automation"
        ]
      }
    },
    "/api/v1/staff/sync": {
      "post": {
        "summary": "Derived POST /api/v1/staff/sync",
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "security": [
          {
            "staffAuth": []
          }
        ],
        "x-csrf-required": true,
        "x-tenant-context": true,
        "x-required-permissions": [
          "staff.sync"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StaffSyncAwardsRequest"
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "staffAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "__Host-pf_staff"
      },
      "customerAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "__Host-pf_customer"
      },
      "superAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "__Host-pf_super"
      }
    },
    "schemas": {
      "StaffLoginRequest": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "maxLength": 255
          },
          "password": {
            "type": "string",
            "minLength": 1,
            "maxLength": 128
          }
        },
        "required": [
          "email",
          "password"
        ],
        "additionalProperties": false
      },
      "AwardPointsRequest": {
        "type": "object",
        "properties": {
          "customerQrToken": {
            "type": "string",
            "minLength": 1
          },
          "amount_q": {
            "type": "number",
            "minimum": 0,
            "maximum": 1000000
          },
          "visits": {
            "type": "integer",
            "exclusiveMinimum": 0,
            "maximum": 1000
          },
          "items": {
            "type": "integer",
            "exclusiveMinimum": 0,
            "maximum": 10000
          },
          "meta": {
            "type": "object",
            "additionalProperties": {}
          },
          "txId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "customerQrToken",
          "txId"
        ],
        "additionalProperties": false
      },
      "StaffRedeemRequest": {
        "type": "object",
        "properties": {
          "customerId": {
            "type": "string",
            "format": "uuid"
          },
          "rewardId": {
            "$ref": "#/definitions/StaffRedeemRequest/properties/customerId"
          },
          "requestId": {
            "$ref": "#/definitions/StaffRedeemRequest/properties/customerId"
          }
        },
        "required": [
          "customerId",
          "rewardId",
          "requestId"
        ],
        "additionalProperties": false
      },
      "StaffSyncAwardsRequest": {
        "type": "object",
        "properties": {
          "awards": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "txId": {
                  "type": "string",
                  "format": "uuid"
                },
                "customerQrToken": {
                  "type": "string",
                  "minLength": 20
                },
                "amount_q": {
                  "type": "number",
                  "minimum": 0
                },
                "visits": {
                  "type": "integer",
                  "exclusiveMinimum": 0
                },
                "items": {
                  "type": "integer",
                  "exclusiveMinimum": 0
                },
                "meta": {},
                "client_ts": {
                  "type": "string"
                }
              },
              "required": [
                "txId",
                "customerQrToken"
              ],
              "additionalProperties": false
            },
            "maxItems": 200
          }
        },
        "required": [
          "awards"
        ],
        "additionalProperties": false
      },
      "StaffRefundRequest": {
        "type": "object",
        "properties": {
          "transactionId": {
            "type": "string",
            "format": "uuid"
          },
          "requestId": {
            "type": "string",
            "format": "uuid"
          },
          "reason": {
            "type": "string",
            "minLength": 2,
            "maxLength": 200
          }
        },
        "required": [
          "transactionId",
          "requestId"
        ],
        "additionalProperties": false
      },
...
      "StaffCreateRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 2,
            "maxLength": 120
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "phone": {
            "type": "string"
          },
          "password": {
            "type": "string",
            "minLength": 6,
            "maxLength": 128
          },
          "role": {
            "type": "string",
            "enum": [
              "CASHIER",
              "MANAGER"
            ]
          },
          "branch_id": {
            "type": "string",
            "format": "uuid"
          },
          "can_manage_gift_cards": {
            "type": "boolean"
          }
        },
        "required": [
          "name",
          "email",
          "password"
        ],
        "additionalProperties": false
      },
      "StaffUpdateRequest": {
        "type": "object",
        "properties": {
          "active": {
            "type": "boolean"
          },
          "password": {
            "type": "string",
            "minLength": 6,
            "maxLength": 128
          },
          "role": {
            "type": "string",
            "enum": [
              "CASHIER",
              "MANAGER"
            ]
          },
          "branch_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ]
          },
          "can_manage_gift_cards": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          },
          "code": {
            "type": "string"
          }
        },
        "required": [
          "error"
        ],
        "additionalProperties": false
      },
...
      "StaffListResponse": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean",
            "const": true
          },
          "staff": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string",
                  "format": "uuid"
                },
                "name": {
                  "type": "string"
                },
                "email": {
                  "anyOf": [
                    {
                      "type": "string",
                      "format": "email"
                    },
                    {
                      "type": "null"
                    }
                  ]
                },
                "phone": {
                  "type": [
                    "string",
                    "null"
                  ]
                },
                "role": {
                  "type": "string"
                },
                "branch_id": {
                  "anyOf": [
                    {
                      "type": "string",
                      "format": "uuid"
                    },
                    {
                      "type": "null"
                    }
                  ]
                },
                "active": {
                  "type": "boolean"
                },
                "can_manage_gift_cards": {
                  "type": "boolean"
                },
                "created_at": {
                  "type": "string"
                }
              },
              "required": [
                "id",
                "name",
                "role"
              ],
              "additionalProperties": true
            }
          }
        },
        "required": [
          "ok",
          "staff"
        ],
        "additionalProperties": false
      },
      "StaffDetailResponse": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean",
            "const": true
          },
          "staff": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string",
                "format": "uuid"
              },
              "name": {
                "type": "string"
              },
              "email": {
                "anyOf": [
                  {
                    "type": "string",
                    "format": "email"
                  },
                  {
                    "type": "null"
                  }
                ]
              },
              "phone": {
                "type": [
                  "string",
                  "null"
                ]
              },
              "role": {
                "type": "string"
              },
              "branch_id": {
                "anyOf": [
                  {
                    "type": "string",
                    "format": "uuid"
                  },
                  {
                    "type": "null"
                  }
                ]
              },
              "active": {
                "type": "boolean"
              },
              "can_manage_gift_cards": {
                "type": "boolean"
              },
              "created_at": {
                "type": "string"
              }
            },
            "required": [
              "id",
              "name",
              "role"
            ],
            "additionalProperties": true
          }
        },
        "required": [
          "ok",
          "staff"
        ],
        "additionalProperties": false
      },
...
      "StaffAwardResponse": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean",
            "const": true
          },
          "pointsAwarded": {
            "type": "number"
          },
          "newBalance": {
            "type": "number"
          },
          "newPendingBalance": {
            "type": "number"
          },
          "customerId": {
            "type": "string",
            "format": "uuid"
          },
          "transactionId": {
            "type": "string",
            "format": "uuid"
          },
          "status": {
            "type": "string"
          },
          "availableAt": {
            "type": [
              "string",
              "null"
            ]
          }
        },
        "required": [
          "ok",
          "pointsAwarded",
          "newBalance",
          "customerId",
          "transactionId"
        ],
        "additionalProperties": true
      },
      "StaffRefundResponse": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean",
            "const": true
          },
          "transactionId": {
            "type": "string",
            "format": "uuid"
          },
          "reversalTransactionId": {
            "type": "string",
            "format": "uuid"
          },
          "customerId": {
            "type": "string",
            "format": "uuid"
          },
          "pointsEffect": {
            "type": "number"
          },
          "gamificationReconciliation": {
            "type": "object",
            "properties": {
              "achievementsRevoked": {
                "type": "number"
              },
              "challengesRevoked": {
                "type": "number"
              }
            },
            "required": [
              "achievementsRevoked",
              "challengesRevoked"
            ],
            "additionalProperties": false
          },
          "newBalance": {
            "type": "number"
          },
          "newPendingBalance": {
            "type": "number"
          }
        },
        "required": [
          "ok",
          "transactionId",
          "reversalTransactionId",
          "customerId",
          "pointsEffect",
          "gamificationReconciliation",
          "newBalance",
          "newPendingBalance"
        ],
        "additionalProperties": true
      }
    }
  }
}